# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Security Toolkit API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /speed/test:
    post:
      operationId: runSpeedTest
      tags: [speed]
      summary: Run page speed test
      description: Measure page load performance metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: Speed test results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeedTestResult"
  /seo/audit:
    post:
      operationId: runSeoAudit
      tags: [seo]
      summary: Run SEO audit
      description: Analyze page for SEO issues and opportunities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanRequest"
      responses:
        "200":
          description: SEO audit results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeoAuditResult"
...
components:
  schemas:
...
    ScanRequest:
      type: object
      properties:
        url:
          type: string
          description: Target URL to scan
      required:
        - url
...
    SpeedMetric:
      type: object
      properties:
        name:
          type: string
        value:
          type: number
        unit:
          type: string
        score:
          type: integer
        rating:
          type: string
          enum: [good, needs_improvement, poor]
      required:
        - name
        - value
        - unit
        - score
        - rating
    SpeedTestResult:
      type: object
      properties:
        url:
          type: string
        overallScore:
          type: integer
        grade:
          type: string
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/SpeedMetric"
        resourceSummary:
          type: object
          properties:
            totalRequests:
              type: integer
            totalSize:
              type: integer
            htmlSize:
              type: integer
            cssSize:
              type: integer
            jsSize:
              type: integer
            imageSize:
              type: integer
          required:
            - totalRequests
            - totalSize
        testedAt:
          type: string
      required:
        - url
        - overallScore
        - grade
        - metrics
        - resourceSummary
        - testedAt
    SeoIssue:
      type: object
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [error, warning, info]
        message:
          type: string
        element:
          type: string
      required:
        - type
        - severity
        - message
    SeoAuditResult:
      type: object
      properties:
        url:
          type: string
        score:
          type: integer
        grade:
          type: string
        title:
          type: string
        description:
          type: string
        canonicalUrl:
          type: string
        robotsMeta:
          type: string
        openGraph:
          type: object
          additionalProperties: true
        twitterCard:
          type: object
          additionalProperties: true
        structuredData:
          type: array
          items:
            type: object
            additionalProperties: true
        headings:
          type: object
          additionalProperties: true
        issues:
          type: array
          items:
            $ref: "#/components/schemas/SeoIssue"
        internalLinks:
          type: integer
        externalLinks:
          type: integer
        brokenLinks:
          type: integer
        imagesWithoutAlt:
          type: integer
        scannedAt:
          type: string
      required:
        - url
        - score
        - grade
        - issues
        - scannedAt
...