# lib/api-zod/src/generated/types/seoAuditResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoAuditResultHeadings } from "./seoAuditResultHeadings";
import type { SeoAuditResultOpenGraph } from "./seoAuditResultOpenGraph";
import type { SeoAuditResultStructuredDataItem } from "./seoAuditResultStructuredDataItem";
import type { SeoAuditResultTwitterCard } from "./seoAuditResultTwitterCard";
import type { SeoIssue } from "./seoIssue";

export interface SeoAuditResult {
  url: string;
  score: number;
  grade: string;
  title?: string;
  description?: string;
  canonicalUrl?: string;
  robotsMeta?: string;
  openGraph?: SeoAuditResultOpenGraph;
  twitterCard?: SeoAuditResultTwitterCard;
  structuredData?: SeoAuditResultStructuredDataItem[];
  headings?: SeoAuditResultHeadings;
  issues: SeoIssue[];
  internalLinks?: number;
  externalLinks?: number;
  brokenLinks?: number;
  imagesWithoutAlt?: number;
  scannedAt: string;
}