# lib/api-zod/src/generated/types/seoIssue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SeoIssueSeverity } from "./seoIssueSeverity";

export interface SeoIssue {
  type: string;
  severity: SeoIssueSeverity;
  message: string;
  element?: string;
}