# lib/api-zod/src/generated/types/speedTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Security Toolkit API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SpeedMetric } from "./speedMetric";
import type { SpeedTestResultResourceSummary } from "./speedTestResultResourceSummary";

export interface SpeedTestResult {
  url: string;
  overallScore: number;
  grade: string;
  metrics: SpeedMetric[];
  resourceSummary: SpeedTestResultResourceSummary;
  testedAt: string;
}